Split ConfigurationState into 3 parts: Migration, Model, and everything else.
